Code cleanup in toolchain tester
