+++ tools/toolchain-tester/printer.py
@@ -1,13 +1,11 @@
 # Prevents an import at runtime so there's no cyclical dependency
 from __future__ import annotations
 from typing import TYPE_CHECKING
+from typing import List, Optional, Tuple
+from errors import TestFailure
 
 if TYPE_CHECKING:
     from tests import Test
-
-from typing import List, Optional, Tuple
-
-from settings import TestFailure
 
 COLORS = {
     "black": 0,
@@ -64,9 +62,10 @@ def print_test_results(
 
 
 def print_test_results_machine(
-    results: List[Tuple[Test, Optional[TestFailure]]], run_time: float
+    _results: List[Tuple[Test, Optional[TestFailure]]], _run_time: float
 ) -> None:
     Printer.red("TODO")
+    return []
 
 
 class Printer:
