+++ tools/toolchain-tester/testrunner.py
@@ -1,9 +1,9 @@
 from multiprocessing import Pool
-from typing import Dict, List, NamedTuple, Optional, Tuple
+from typing import Dict, List, NamedTuple, Optional
 
 from printer import Printer as P
 
-from settings import TestFailure
+from errors import TestFailure
 from testsuite import TestSuite
 from tests import Test
 
@@ -23,7 +23,7 @@ class TestRunner:
         self.num_jobs: int = num_jobs
 
         self.test_map: Dict[str, Test] = {}
-        self.test_suite_map: Dict[str, TestSuite] = {}
+        self.test_suites_map: Dict[str, TestSuite] = {}
 
     def get_tests_to_run(self):
         tests_to_run: List[Test] = []
@@ -38,7 +38,7 @@ class TestRunner:
             # We're running a test suite
             self.build_test_suite_map()
             ts = self.test_suites_map[self.name_of_tests_to_run]
-            tests_to_run = [t for t in ts.tests]
+            tests_to_run = ts.tests
 
         return tests_to_run
 
@@ -58,12 +58,12 @@ class TestRunner:
         return TestResult(t, None)
 
     def build_test_suite_map(self):
-        test_suites_map = {}
+        tst_suites_map = {}
 
         for ts in self.test_suites:
-            test_suites_map[ts.name] = ts
+            tst_suites_map[ts.name] = ts
 
-        self.test_suites_map = test_suites_map
+        self.test_suites_map = tst_suites_map
 
     def build_test_map(self):
         test_map = {}
